requirements.txt
langchain
langchain-community
langchain-openai
pypdf
python-dotenv
chromadb
tiktoken